Complete lab05

+++ Lab02/AimsProject/src/hust/soict/ict/aims/media/Book.java
@@ -10,21 +10,21 @@ public class Book extends Media{
     private ArrayList<String>  listOfAuthors  = new ArrayList<>();
 
 
-    public Book(int id, String title, String category, String author, double cost, int token){
+    public Book(int id, String title, String category, String author, float cost, int token){
         super(id, title,category, cost);
         this.token = token;
         this.listOfAuthors.add(author);
     }
-    public Book(int id, String title, String category, double cost, int token){
+    public Book(int id, String title, String category, float cost, int token){
         super(id, title,category, cost);
         this.token = token;
     }
-    public Book(String title, String category, double cost, int token){
+    public Book(String title, String category, float cost, int token){
         super(title,category, cost);
         this.token = token;
 
     }
-    public Book(String title, String category,String author, double cost, int token){
+    public Book(String title, String category,String author, float cost, int token){
         super(title,category, cost);
         this.token = token;
         this.listOfAuthors.add(author);
@@ -33,7 +33,7 @@ public class Book extends Media{
         return token;
     }
     @Override
-    // String title, String category, double cost, int token
+    // String title, String category, float cost, int token
     public void toStringItem(int order){
         System.out.println(order + ".Book - " +  this.getTitle()+ " - " + this.getCategory()  + " - " + this.getToken() + " - " + this.getCost() + " $");
     }
@@ -44,7 +44,7 @@ public class Book extends Media{
                 return;
             }
         }
-        listOfAuthors.add(authorInput); 
+        listOfAuthors.add(authorInput);
     }
     public void removeAuthor(String authorInput){
         for(int i = 0; i < listOfAuthors.size(); i++){
@@ -54,14 +54,15 @@ public class Book extends Media{
             }
         }
         System.out.println("Song not found in TrackList");
-            
+
     }
     public boolean isMatchItem(Book item){
         return this.id == item.id &&
-        this.title.equals(item.getTitle())&& 
-        this.category.equals(item.getCategory()) &&
-        this.cost == item.cost && 
-        this.token == item.token;
+                this.title.equals(item.getTitle())&&
+                this.category.equals(item.getCategory()) &&
+                this.cost == item.cost &&
+                this.token == item.token;
     }
+
 
 }
