+++ Lab02/AimsProject/src/hust/soict/ict/aims/media/DigitalVideoDisc.java
@@ -2,7 +2,7 @@ package hust.soict.ict.aims.media;
 
 
 public class DigitalVideoDisc extends Disc implements Playable{
-    public DigitalVideoDisc(String title, String category,String director,int length, double cost) {
+    public DigitalVideoDisc(String title, String category,String director,int length, float cost) {
         super(title,category,director,length,cost);
     }
     public void play(){
@@ -18,11 +18,11 @@ public class DigitalVideoDisc extends Disc implements Playable{
     }
     public boolean isMatchItem(DigitalVideoDisc item){
         return this.id == item.id &&
-        this.title.equals(item.getTitle())&& 
-        this.category.equals(item.getCategory()) &&
-        this.cost == item.cost && 
-        this.director.equals(item.getDirector()) && 
-        this.length == item.length;
+                this.title.equals(item.getTitle())&&
+                this.category.equals(item.getCategory()) &&
+                this.cost == item.cost &&
+                this.director.equals(item.getDirector()) &&
+                this.length == item.length;
     }
 
 }
