+++ Lab02/AimsProject/src/hust/soict/ict/aims/media/Media.java
@@ -6,18 +6,18 @@ public class Media implements Comparable<Media> {
     protected int id;
     protected String title;
     protected String category;
-    protected double cost;
+    protected float cost;
 
-    public Media(int id, String title, String category, double cost){
+    public Media(int id, String title, String category, float cost){
         this.id = id;
         this.title = title;
         this.category = category;
-        this.cost = cost; 
+        this.cost = cost;
     }
-    public Media(String title, String category, double cost){
+    public Media(String title, String category, float  cost){
         this.title = title;
         this.category = category;
-        this.cost = cost; 
+        this.cost = cost;
     }
 
 
@@ -30,7 +30,7 @@ public class Media implements Comparable<Media> {
     public String getCategory() {
         return category;
     }
-    public double getCost() {
+    public float getCost() {
         return cost;
     }
     public void toStringItem(int order) {
@@ -38,9 +38,9 @@ public class Media implements Comparable<Media> {
     }
     public boolean isMatchItem(Media item){
         return this.id == item.id &&
-            this.title == item.title && 
-            this.category == item.category &&
-            this.cost == item.cost; 
+                this.title == item.title &&
+                this.category == item.category &&
+                this.cost == item.cost;
     }
     @Override
     public boolean equals(Object o) {
@@ -54,7 +54,7 @@ public class Media implements Comparable<Media> {
         // Sắp xếp theo tiêu đề sau đó giá
         int result = this.title.compareTo(other.title);
         if (result == 0) {
-            result = Double.compare(other.cost, this.cost);
+            result = Float.compare(other.cost, this.cost);
         }
         return result;
     }
@@ -63,7 +63,7 @@ public class Media implements Comparable<Media> {
     public static Comparator<Media> COMPARE_BY_COST_TITLE = new Comparator<Media>() {
         @Override
         public int compare(Media media1, Media media2) {
-            int result = Double.compare(media2.cost, media1.cost);
+            int result = Float.compare(media2.cost, media1.cost);
             if (result == 0) {
                 result = media1.title.compareTo(media2.title);
             }
@@ -77,10 +77,10 @@ public class Media implements Comparable<Media> {
         public int compare(Media media1, Media media2) {
             int result = media1.title.compareTo(media2.title);
             if (result == 0) {
-                result = Double.compare(media2.cost, media1.cost);
+                result = Float.compare(media2.cost, media1.cost);
             }
             return result;
         }
     };
-    
+
 }
